updated delete candidates

+++ src/applications/application.service.ts
@@ -7,6 +7,7 @@ import { ProjectApplication } from './application.entity';
 import { UserAccounts } from 'src/auth/User.entity';
 import { AccountProject } from 'src/admin/projects/project.entity';
 import { RecruiterProject } from 'src/recruiter/projects/project.entity';
+import { RecruiterCompanyUser } from 'src/recruiter/recruiter-company/recruiter-company-user.entity';
 @Injectable()
 export class ApplicationService {
   constructor(
@@ -15,7 +16,9 @@ export class ApplicationService {
     @InjectRepository(UserAccounts)
     private readonly userRepository: Repository<UserAccounts>,
     @InjectRepository(RecruiterProject)
-    private readonly projectRepository: Repository<RecruiterProject>
+    private readonly projectRepository: Repository<RecruiterProject>,
+    @InjectRepository(RecruiterCompanyUser)
+    private recruiterCompanyUserRepository: Repository<RecruiterCompanyUser>,
   ) {}
 
   async createApplication(body: any, userId: number): Promise<any> {
@@ -72,7 +75,7 @@ export class ApplicationService {
     }
   }
 
-  async deleteApplicationsForUserAndCompany(userId: number, companyId: number, loggedInUser: number): Promise<any> {
+  async deleteApplicationsForUserAndCompany(userId: number, loggedInUser: number): Promise<any> {
     try {
       // Check if the logged-in user is an Admin
       const checkAdmin = await this.userRepository.findOne({
@@ -82,13 +85,22 @@ export class ApplicationService {
       if (!checkAdmin) {
         return { error: true, message: 'You are not an admin User.' };
       }
+
+      const recruiterCompanyUser = await this.recruiterCompanyUserRepository.findOne({
+        where: { user: { id: userId } },
+        relations: ['company'],
+      });
+      
+      if (!recruiterCompanyUser) {
+        return { error: true, message: 'User is not associated with any recruiter company.' };
+      }
   
       // Delete applications for the specified user and company
       await this.applicationRepository.createQueryBuilder()
         .delete()
         .from(ProjectApplication)
         .where('userId = :userId', { userId })
-        .andWhere('projectId IN (SELECT id FROM recruiter_project WHERE companyId = :companyId)', { companyId })
+        .andWhere('projectId IN (SELECT id FROM recruiter_project WHERE companyId = :companyId)', { companyId:recruiterCompanyUser.company.id })
         .execute();
   
       return { error: false, message: 'Applications deleted successfully.' };
